+++ js/sprite/Title.js
@@ -55,7 +55,7 @@ export default class Title{
 
             this.newPosX = this.suposedX + this.x;
             this.newPosY = this.suposedY + this.y;
-            let newAnimKey = [[parseInt(this.node.style.left),parseInt(this.node.style.top), 0, 0], [this.newPosX, this.newPosY, 10, 0.1]];
+            let newAnimKey = [[parseFloat(this.node.style.left),parseFloat(this.node.style.top), 0, 0], [this.newPosX, this.newPosY, 10, 0.1]];
             this.anim = null;
             this.anim = new AnimationHelper(this.node, newAnimKey);
         }
